Specify __constant for OpenCL kernel
The __constant space is a memory region where work-items have faster
access to the data. The limitation, of course, is that OpenCL kernels
can only take a limited number of __constant arguments and buffers
stored in the __constant space have limited size. For blake2b case,
a set of the kernel arguments would not exceed 64 KB, and there is
no reason not to allocate them in __constant space.